+++ src/runtime_config/default_whitelists.h
@@ -26,7 +26,14 @@ namespace ccf
       Tables::SERVICE}},
 
     {MEMBER_CAN_PROPOSE,
-     {Tables::WHITELISTS, Tables::GOV_SCRIPTS, Tables::APP_SCRIPTS}},
+     {Tables::MEMBERS,
+      Tables::MEMBER_CERTS,
+      Tables::USERS,
+      Tables::USER_CERTS,
+      Tables::VALUES,
+      Tables::WHITELISTS,
+      Tables::GOV_SCRIPTS,
+      Tables::APP_SCRIPTS}},
 
     {USER_APP_CAN_READ_ONLY,
      {Tables::MEMBERS,
